move singleton_cache.py into singleton
